+++ tests/LineTest.php
@@ -22,10 +22,10 @@ class LineTest extends TestCase
         }));
 
         $this->assertNull($line->while(function (&$index, Line $line) {
-            ++$index;
+            $index++;
         }));
         $this->assertNull($line->while(function (&$index, Line $line) {
-            --$index;
+            $index--;
         }));
         $this->assertFalse($line->previous());
     }
@@ -37,5 +37,43 @@ class LineTest extends TestCase
         $this->expectException('\Exception');
         $this->expectExceptionMessage('Unable to find a next element. Invalid DELTA on \'<strong>second</strong>\'. Maybe your delta code does not end with a newline?');
         $response = $lexer->render();
+    }
+
+    public function testWhileNext()
+    {
+        $lexer = new Lexer('[
+            {
+                "insert": "line 1\nline 2\nline 3\n"
+            }
+        ]');
+        $lexer->render();
+
+        $line0 = $lexer->getLine(0);
+
+        $this->assertSame('line 1', $line0->input);
+
+        $line1 = null;
+        $line0->whileNext(function($line) use (&$line1) {
+            $line1 = $line;
+            return false;
+        });
+        
+        $this->assertSame('line 2', $line1->input);
+        
+        $prevLine = null;
+        $line1->whilePrevious(function($line) use (&$prevLine) {
+            $prevLine = $line;
+            return false;
+        });
+
+        $this->assertSame('line 1', $prevLine->input);
+
+        $value = false;
+        $lineNoPrevious = $line0->whilePrevious(function($line) use (&$value) {
+            $value = true;
+        });
+
+        // should never change to true as there is no previous element.
+        $this->assertFalse($value);
     }
 }